chore: rename directory `core` to `modules`
